move sysprobe config fetcher into subpackage
