+++ comp/metadata/inventoryagent/inventoryagentimpl/inventoryagent_test.go
@@ -25,6 +25,7 @@ import (
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig"
 	"github.com/DataDog/datadog-agent/comp/core/sysprobeconfig/sysprobeconfigimpl"
 	configFetcher "github.com/DataDog/datadog-agent/pkg/config/fetcher"
+	sysprobeConfigFetcher "github.com/DataDog/datadog-agent/pkg/config/fetcher/sysprobe"
 	pkgconfigmodel "github.com/DataDog/datadog-agent/pkg/config/model"
 	pkgconfigsetup "github.com/DataDog/datadog-agent/pkg/config/setup"
 	"github.com/DataDog/datadog-agent/pkg/serializer"
@@ -456,7 +457,7 @@ func TestFetchSystemProbeAgent(t *testing.T) {
 	}
 
 	defer func() {
-		fetchSystemProbeConfig = configFetcher.SystemProbeConfig
+		fetchSystemProbeConfig = sysprobeConfigFetcher.SystemProbeConfig
 	}()
 	fetchSystemProbeConfig = func(config pkgconfigmodel.Reader) (string, error) {
 		// test that the system-probe config was passed and not the agent config
