+++ pkg/config/fetcher/from_processes.go
@@ -10,7 +10,6 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/DataDog/datadog-agent/cmd/system-probe/api/client"
 	"github.com/DataDog/datadog-agent/comp/core/config"
 	"github.com/DataDog/datadog-agent/pkg/api/util"
 	settingshttp "github.com/DataDog/datadog-agent/pkg/config/settings/http"
@@ -106,20 +105,4 @@ func ProcessAgentConfig(config config.Reader, getEntireConfig bool) (string, err
 	client := settingshttp.NewClient(c, ipcAddressWithPort, "process-agent", settingshttp.NewHTTPClientOptions(util.CloseConnection))
 
 	return client.FullConfig()
-}
-
-// SystemProbeConfig fetch the configuration from the system-probe process by querying its API
-func SystemProbeConfig(config config.Reader) (string, error) {
-	hc := client.Get(config.GetString("system_probe_config.sysprobe_socket"))
-
-	c := settingshttp.NewClient(hc, "http://localhost/config", "system-probe", settingshttp.NewHTTPClientOptions(util.CloseConnection))
-	return c.FullConfig()
-}
-
-// SystemProbeConfigBySource fetch the all configuration layers from the system-probe process by querying its API
-func SystemProbeConfigBySource(config config.Reader) (string, error) {
-	hc := client.Get(config.GetString("system_probe_config.sysprobe_socket"))
-
-	c := settingshttp.NewClient(hc, "http://localhost/config", "system-probe", settingshttp.NewHTTPClientOptions(util.CloseConnection))
-	return c.FullConfigBySource()
 }
